+++ packages/amplify-codegen/src/utils/downloadIntrospectionSchema.js
@@ -16,7 +16,11 @@ async function downloadIntrospectionSchema(context, apiId, downloadLocation, reg
       const introspectionDir = dirname(downloadLocation);
       fs.ensureDirSync(introspectionDir);
       fs.writeFileSync(downloadLocation, schema, 'utf8');
-      return relative(amplify.getEnvInfo().projectPath, downloadLocation);
+      try {
+        return relative(amplify.getEnvInfo().projectPath, downloadLocation);
+      } catch {
+        return downloadLocation;
+      }
     } catch (ex) {
       if (ex.code === 'NotFoundException') {
         throw new AmplifyCodeGenAPINotFoundError(constants.ERROR_APPSYNC_API_NOT_FOUND);