+++ packages/graphql-generator/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/aws-amplify/amplify-codegen/compare/@aws-amplify/graphql-generator@0.1.0...@aws-amplify/graphql-generator@0.1.1) (2023-09-19)
+
+**Note:** Version bump only for package @aws-amplify/graphql-generator
+
 # 0.1.0 (2023-09-12)
 
 ### Features